+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 2.0.0.dev3 (unreleased)
 -----------------------
 
+- Handle contentrules by the plone events
+  [folix-01]
 - Change "day" type in week_table (TODO: need an upgrade step?)
   [mauro]